TMC2209 are limited to 1.2 Amps
Also, "RMS" vs. "Peak" is really not a meaningful distinction.
Motors are rated as max amps per phase, which, since the control
signal is DC, is obviously peak and RMS (RMS = Peak).  In cases
where peak is higher than RMS, that means we are varying the current
and for current limits, RMS is what controls them... and RMS is
defined such that it describes the effective limit.

For example a 1A DC through a resistor will heat it exactly as much
as 1A RMS (AC or DC) through the same resistor in *every* case.
For a sine wave, the peak will be 1.414, bit for DC the "peak" is 1A.
Since the primary controling issue in steppers is the resistive
heating, we can treat it like a resistor and just treat the 1A per
phase rating as RMS because we're allowed to put 1A DC through each
phase.